fixing pop/sub
pop is not used by AddTest, so it is not under tests
SP looks at the next empty space, so need to manipulate
SP BEFORE accessing values

sub is x-y, where x was first in the stack.
Switch order of operands to support that.
The other implemented commands are commutative and don't have this bug